About to further prettify checkoutsheet (adding borders and bg colors for debugging spacing)
