+++ Tappayo/ItemEditorView.swift
@@ -16,15 +16,6 @@ struct ItemEditorView: View {
     @State private var editedName: String = ""
     @State private var editedPriceInCents: Int
     @State private var currentQuantity: Int
-    @FocusState private var isNameFieldFocused: Bool
-    @FocusState private var isPriceFieldFocused: Bool
-
-    // Icon editing state
-    @State private var showingIconPicker = false
-    @State private var showingCamera = false
-    @State private var showingPhotoLibrary = false
-    @State private var selectedImage: UIImage?
-    @State private var showingEmojiPicker = false
 
     init(item: CartItem, basketIndex: Int, basket: Binding<[CartItem]>, savedProducts: Binding<[Product]>, formatAmount: @escaping (Int, Bool) -> String) {
         self.item = item
@@ -61,67 +52,22 @@ struct ItemEditorView: View {
         VStack(spacing: 24) {
             Spacer()
 
-            // Product icon display (tappable to edit)
-            if let product = currentProduct {
-                Button(action: {
-                    showingIconPicker = true
-                }) {
-                    VStack {
-                        if let photoFilename = product.photoFilename,
-                           let image = PhotoStorageHelper.loadPhoto(photoFilename) {
-                            // Show photo
-                            Image(uiImage: image)
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 100, height: 100)
-                                .clipShape(RoundedRectangle(cornerRadius: 12))
-                        } else if let emoji = product.emoji {
-                            // Show emoji
-                            Text(emoji)
-                                .font(.system(size: 80))
-                                .frame(width: 100, height: 100)
-                        } else {
-                            // Show placeholder
-                            Image(systemName: "camera.circle")
-                                .font(.system(size: 80))
-                                .foregroundColor(.gray)
-                                .frame(width: 100, height: 100)
-                        }
-                    }
-                }
-                .buttonStyle(.plain)
-                .padding(.bottom, 8)
+            // Product icon editing (only for saved products)
+            if let product = currentProduct,
+               let productIndex = savedProducts.firstIndex(where: { $0.id == product.id }) {
+                ProductIconPicker(product: $savedProducts[productIndex], savedProducts: $savedProducts)
+                    .padding(.bottom, 8)
             }
 
-            // Name editing section
-            VStack(alignment: .leading, spacing: 8) {
-                Text(item.isProduct ? "Product Name" : "Item Name")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-                TextField("Enter name", text: $editedName)
-                    .font(.title3)
-                    .textFieldStyle(.roundedBorder)
-                    .focused($isNameFieldFocused)
-                    .onSubmit {
-                        saveChanges()
-                    }
-            }
-            .padding(.horizontal)
+            // Name editing
+            ProductNameField(
+                name: $editedName,
+                label: item.isProduct ? "Product Name" : "Item Name",
+                onSubmit: saveChanges
+            )
 
-            // Price editing section
-            VStack(alignment: .leading, spacing: 8) {
-                Text("Price")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-                CurrencyTextField(
-                    value: $editedPriceInCents,
-                    placeholder: "$0.00",
-                    font: .title3
-                )
-                .frame(height: 40)
-                .focused($isPriceFieldFocused)
-            }
-            .padding(.horizontal)
+            // Price editing
+            ProductPriceField(priceInCents: $editedPriceInCents)
 
             // Calculation row (only if quantity > 1)
             if currentQuantity > 1 {
@@ -196,38 +142,6 @@ struct ItemEditorView: View {
             .padding(.bottom)
         }
         .background(Color(.systemBackground))
-        .confirmationDialog("Choose Icon", isPresented: $showingIconPicker) {
-            Button("Choose Emoji") {
-                showingEmojiPicker = true
-            }
-            Button("Take Photo") {
-                showingCamera = true
-            }
-            Button("Choose from Library") {
-                showingPhotoLibrary = true
-            }
-            Button("Cancel", role: .cancel) {}
-        }
-        .sheet(isPresented: $showingCamera) {
-            ImagePicker(image: $selectedImage, sourceType: .camera)
-        }
-        .sheet(isPresented: $showingPhotoLibrary) {
-            ImagePicker(image: $selectedImage, sourceType: .photoLibrary)
-        }
-        .onChange(of: selectedImage) { newImage in
-            handleImageSelection(newImage)
-        }
-        .emojiPicker(
-            isPresented: $showingEmojiPicker,
-            selectedEmoji: Binding(
-                get: {
-                    currentProduct?.emoji ?? ""
-                },
-                set: { newEmoji in
-                    handleEmojiSelection(newEmoji)
-                }
-            )
-        )
         .onDisappear {
             // Auto-save changes when sheet is dismissed (swipe, Done button, or any dismissal method)
             // This prevents data loss if user swipes down without tapping Done
@@ -264,48 +178,6 @@ struct ItemEditorView: View {
             isProduct: item.isProduct
         )
         basket[basketIndex] = updatedItem
-    }
-
-    private func handleImageSelection(_ image: UIImage?) {
-        guard let image = image,
-              item.isProduct,
-              let productId = item.productId,
-              let productIndex = savedProducts.firstIndex(where: { $0.id == productId }) else {
-            selectedImage = nil
-            return
-        }
-
-        // Save the new photo
-        if let filename = PhotoStorageHelper.savePhoto(image) {
-            // Delete old photo if exists
-            if let oldFilename = savedProducts[productIndex].photoFilename {
-                PhotoStorageHelper.deletePhoto(oldFilename)
-            }
-            // Set new photo and clear emoji (photo takes priority)
-            savedProducts[productIndex].photoFilename = filename
-            savedProducts[productIndex].emoji = nil
-            UserDefaults.standard.savedProducts = savedProducts
-        }
-
-        selectedImage = nil
-    }
-
-    private func handleEmojiSelection(_ emoji: String) {
-        guard item.isProduct,
-              let productId = item.productId,
-              let productIndex = savedProducts.firstIndex(where: { $0.id == productId }) else {
-            return
-        }
-
-        // Delete old photo if exists
-        if let oldFilename = savedProducts[productIndex].photoFilename {
-            PhotoStorageHelper.deletePhoto(oldFilename)
-        }
-
-        // Set emoji and clear photo
-        savedProducts[productIndex].emoji = emoji.isEmpty ? nil : emoji
-        savedProducts[productIndex].photoFilename = nil
-        UserDefaults.standard.savedProducts = savedProducts
     }
 }
